requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
billiard==3.6.4.0
braintree==4.14.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.2
django-allauth==0.49.0
django-autoslug==1.9.8
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-import-export==2.8.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.2.3
MarkupPy==1.14
msgpack==1.0.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stripe==2.74.0
tablib==3.2.1
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
